refactor(agent): extract startup skills bootstrap module (#232)
